Move AppEnvironment to templates
