+++ gatsby-browser.js
@@ -4,7 +4,7 @@ import 'prismjs/themes/prism-okaidia.css';
 
 import React from 'react';
 
-import AppEnvironment from 'components/atoms/AppEnvironment';
+import AppEnvironment from 'templates/AppEnvironment';
 
 export function wrapRootElement({ element }) {
   return (
